requirements.txt
google-generativeai==0.8.3
fastapi[standard]
pymongo==4.10.1